+++ firmware/lib/core/util.cpp
@@ -24,7 +24,7 @@ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
 
 #include <util.h>
 
-namespace OxUtil {
+namespace OxCore {
 
 bool CheckArrayBounds(int index, int length) {
     return (index >= 0 && index < length);
